src/main/webapp/app/entities/cso/delete/cso-delete-dialog.component.ts
import { Component } from '@angular/core';
import { FormsModule } from '@angular/forms';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import SharedModule from 'app/shared/shared.module';
import { ICso } from '../cso.model';
import { CsoService } from '../service/cso.service';
import { ITEM_DELETED_EVENT } from 'app/config/navigation.constants';

@Component({
  standalone: true,
  templateUrl: './cso-delete-dialog.component.html',
  imports: [SharedModule, FormsModule],
})
export class CsoDeleteDialogComponent {
  cso?: ICso;

  constructor(protected csoService: CsoService, protected activeModal: NgbActiveModal) {}

  cancel(): void {
    this.activeModal.dismiss();
  }

  confirmDelete(id: number): void {
    this.csoService.delete(id).subscribe(() => {
      this.activeModal.close(ITEM_DELETED_EVENT);
    });
  }
}
